Extract SparkBulkImportDriver, SparkFileInfoRow
