+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/job/runner/BulkImportJobRunner.java
@@ -24,48 +24,25 @@ import com.amazonaws.services.securitytoken.AWSSecurityTokenServiceClientBuilder
 import com.amazonaws.services.securitytoken.model.GetCallerIdentityRequest;
 import com.amazonaws.services.securitytoken.model.GetCallerIdentityResult;
 import com.google.gson.JsonSyntaxException;
-import com.joom.spark.ExplicitRepartitionStrategy$;
-import org.apache.hadoop.conf.Configuration;
-import org.apache.spark.SparkConf;
-import org.apache.spark.SparkContext;
-import org.apache.spark.api.java.JavaSparkContext;
-import org.apache.spark.broadcast.Broadcast;
-import org.apache.spark.serializer.KryoSerializer;
 import org.apache.spark.sql.Dataset;
 import org.apache.spark.sql.Row;
-import org.apache.spark.sql.SparkSession;
-import org.apache.spark.sql.execution.SparkStrategy;
-import org.apache.spark.sql.types.DataTypes;
-import org.apache.spark.sql.types.StructType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import scala.collection.JavaConverters;
 
 import sleeper.bulkimport.job.BulkImportJob;
 import sleeper.bulkimport.job.BulkImportJobSerDe;
 import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.table.TableProperties;
 import sleeper.configuration.properties.table.TablePropertiesProvider;
-import sleeper.core.partition.Partition;
-import sleeper.core.schema.Schema;
 import sleeper.statestore.FileInfo;
 import sleeper.statestore.StateStore;
-import sleeper.statestore.StateStoreException;
 import sleeper.statestore.StateStoreProvider;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.nio.file.attribute.PosixFileAttributes;
-import java.time.Duration;
-import java.time.Instant;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
 
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.BULK_IMPORT_BUCKET;
-import static sleeper.configuration.properties.UserDefinedInstanceProperty.FILE_SYSTEM;
 
 /**
  * This abstract class executes a Spark job that reads in input Parquet files and writes
@@ -77,17 +54,13 @@ import static sleeper.configuration.properties.UserDefinedInstanceProperty.FILE_
  */
 public class BulkImportJobRunner {
     private static final Logger LOGGER = LoggerFactory.getLogger(BulkImportJobRunner.class);
-    public static final String PARTITION_FIELD_NAME = "__partition";
-    public static final String FILENAME_FIELD_NAME = "__fileName";
-    public static final String NUM_RECORDS_FIELD_NAME = "__numRecords";
 
-    private final BulkImportPartitioner partitioner;
-
+    private final SparkRecordPartitioner partitioner;
     private InstanceProperties instanceProperties;
     private TablePropertiesProvider tablePropertiesProvider;
     private StateStoreProvider stateStoreProvider;
 
-    public BulkImportJobRunner(BulkImportPartitioner partitioner) {
+    public BulkImportJobRunner(SparkRecordPartitioner partitioner) {
         this.partitioner = partitioner;
     }
 
@@ -98,103 +71,11 @@ public class BulkImportJobRunner {
     }
 
     public void run(BulkImportJob job) throws IOException {
-        Instant startTime = Instant.now();
-        LOGGER.info("Received bulk import job with id {} at time {}", job.getId(), startTime);
-        LOGGER.info("Job is {}", job);
-
-        // Initialise Spark
-        LOGGER.info("Initialising Spark");
-        SparkConf sparkConf = new SparkConf();
-        sparkConf.set("spark.serializer", KryoSerializer.class.getName());
-        sparkConf.registerKryoClasses(new Class[]{Partition.class});
-        SparkSession session = new SparkSession.Builder().config(sparkConf).getOrCreate();
-        scala.collection.immutable.List<SparkStrategy> strategies = JavaConverters.collectionAsScalaIterable(Collections.singletonList((org.apache.spark.sql.execution.SparkStrategy) ExplicitRepartitionStrategy$.MODULE$)).toList();
-        session.experimental().extraStrategies_$eq(strategies);
-        SparkContext sparkContext = session.sparkContext();
-        JavaSparkContext javaSparkContext = JavaSparkContext.fromSparkContext(sparkContext);
-        LOGGER.info("Spark initialised");
-
-        // Load table information
-        LOGGER.info("Loading table properties and schema for table {}", job.getTableName());
-        TableProperties tableProperties = tablePropertiesProvider.getTableProperties(job.getTableName());
-        Schema schema = tableProperties.getSchema();
-        StructType convertedSchema = new StructTypeFactory().getStructType(schema);
-
-        // Load statestore and partitions
-        LOGGER.info("Loading statestore and partitions");
-        StateStore stateStore = stateStoreProvider.getStateStore(tableProperties);
-        List<Partition> allPartitions;
-        try {
-            allPartitions = stateStore.getAllPartitions();
-        } catch (StateStoreException e) {
-            throw new RuntimeException("Failed to load statestore. Are permissions correct for this service account?");
-        }
-
-        Configuration conf = sparkContext.hadoopConfiguration();
-        Broadcast<List<Partition>> broadcastedPartitions = javaSparkContext.broadcast(allPartitions);
-        LOGGER.info("Starting data processing");
-
-        // Create paths to be read
-        List<String> pathsWithFs = new ArrayList<>();
-        String fs = instanceProperties.get(FILE_SYSTEM);
-        LOGGER.info("Using file system {}", fs);
-        job.getFiles().forEach(file -> pathsWithFs.add(fs + file));
-        LOGGER.info("Paths to be read are {}", pathsWithFs);
-
-        // Run bulk import
-        Dataset<Row> dataWithPartition = session.read()
-                .schema(convertedSchema)
-                .option("pathGlobFilter", "*.parquet")
-                .option("recursiveFileLookup", "true")
-                .parquet(pathsWithFs.toArray(new String[0]));
-
-        LOGGER.info("Running bulk import job with id {}", job.getId());
-        List<FileInfo> fileInfos = partitioner.createFileInfos(
-                        SparkPartitionRequest.builder().rows(dataWithPartition)
-                                .instanceProperties(instanceProperties).tableProperties(tableProperties)
-                                .broadcastedPartitions(broadcastedPartitions).conf(conf).build())
-                .collectAsList().stream()
-                .map(this::createFileInfo)
-                .collect(Collectors.toList());
-
-        long numRecords = fileInfos.stream()
-                .mapToLong(FileInfo::getNumberOfRecords)
-                .sum();
-        try {
-            stateStore.addFiles(fileInfos);
-        } catch (StateStoreException e) {
-            throw new RuntimeException("Failed to add files to state store. Ensure this service account has write access. Files may need to "
-                    + "be re-imported for clients to accesss data");
-        }
-        LOGGER.info("Added {} files to statestore", fileInfos.size());
-        Instant finishTime = Instant.now();
-        LOGGER.info("Finished bulk import job {} at time {}", job.getId(), finishTime);
-        long durationInSeconds = Duration.between(startTime, finishTime).getSeconds();
-        double rate = numRecords / (double) durationInSeconds;
-        LOGGER.info("Bulk import job {} took {} seconds (rate of {} per second)", job.getId(), durationInSeconds, rate);
-
-        sparkContext.stop(); // Calling this manually stops it potentially timing out after 10 seconds.
-
+        new SparkBulkImportDriver(partitioner, instanceProperties, tablePropertiesProvider, stateStoreProvider)
+                .run(job);
     }
 
-    private FileInfo createFileInfo(Row row) {
-        return FileInfo.builder()
-                .filename(row.getAs(FILENAME_FIELD_NAME))
-                .jobId(null)
-                .fileStatus(FileInfo.FileStatus.ACTIVE)
-                .partitionId(row.getAs(PARTITION_FIELD_NAME))
-                .numberOfRecords(row.getAs(NUM_RECORDS_FIELD_NAME))
-                .build();
-    }
-
-    public static StructType createFileInfoSchema() {
-        return new StructType()
-                .add(PARTITION_FIELD_NAME, DataTypes.StringType)
-                .add(FILENAME_FIELD_NAME, DataTypes.StringType)
-                .add(NUM_RECORDS_FIELD_NAME, DataTypes.LongType);
-    }
-
-    public static void start(String[] args, BulkImportPartitioner partitioner) throws Exception {
+    public static void start(String[] args, SparkRecordPartitioner partitioner) throws Exception {
         if (args.length != 2) {
             throw new IllegalArgumentException("Expected two arguments, the first with the id of the bulk import job," +
                     " the second with the config bucket");
