+++ java/bulk-import/bulk-import-runner/src/main/java/sleeper/bulkimport/job/runner/dataframe/BulkImportJobDataframeRunner.java
@@ -27,6 +27,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import sleeper.bulkimport.job.runner.BulkImportJobRunner;
+import sleeper.bulkimport.job.runner.SparkFileInfoRow;
 import sleeper.bulkimport.job.runner.SparkPartitionRequest;
 import sleeper.bulkimport.job.runner.StructTypeFactory;
 import sleeper.core.partition.Partition;
@@ -38,9 +39,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static sleeper.bulkimport.job.runner.BulkImportJobRunner.PARTITION_FIELD_NAME;
-import static sleeper.bulkimport.job.runner.BulkImportJobRunner.createFileInfoSchema;
-
 /**
  * The {@link BulkImportJobDataframeRunner} is a {@link BulkImportJobRunner} which
  * uses Spark's Dataframe API to efficiently sort and write out the data split by
@@ -48,6 +46,7 @@ import static sleeper.bulkimport.job.runner.BulkImportJobRunner.createFileInfoSc
  */
 public class BulkImportJobDataframeRunner {
     private static final Logger LOGGER = LoggerFactory.getLogger(BulkImportJobDataframeRunner.class);
+    private static final String PARTITION_FIELD_NAME = "__partition";
 
     private BulkImportJobDataframeRunner() {
     }
@@ -87,7 +86,7 @@ public class BulkImportJobDataframeRunner {
                         request.instanceProperties().saveAsString(),
                         request.tableProperties().saveAsString(),
                         request.conf()),
-                RowEncoder.apply(createFileInfoSchema()));
+                RowEncoder.apply(SparkFileInfoRow.createFileInfoSchema()));
     }
 
     private static StructType createEnhancedSchema(StructType convertedSchema) {
